+++ src/datatypes/api-token.ts
@@ -3,10 +3,10 @@ import type { components } from "@/schema/api.js";
 type Schemas = components["schemas"];
 
 // API token resources.
-export type ApiToken = Schemas["ApiToken"];
-export type ApiTokenWithJWT = Schemas["ApiTokenWithJWT"];
+export type AccountApiToken = Schemas["AccountApiToken"];
+export type AccountApiTokenWithJwt = Schemas["AccountApiTokenWithJwt"];
 export type ApiTokenType = Schemas["ApiTokenType"];
-export type CreateApiToken = Schemas["CreateApiToken"];
-export type UpdateApiToken = Schemas["UpdateApiToken"];
+export type CreateAccountApiToken = Schemas["CreateAccountApiToken"];
+export type UpdateAccountApiToken = Schemas["UpdateAccountApiToken"];
 export type TokenExpiration = Schemas["TokenExpiration"];
-export type ApiTokenPage = Schemas["ApiTokenPage"];
+export type AccountApiTokenPage = Schemas["AccountApiTokenPage"];